Address code review findings
- Remove weak from delegate properties on SwiftUI struct views
- Remove dead code: isARParcelFittingAvailable on custom package VM,
  dimensionsLabel on results VM, unused imports
- Move ARPackageContext to standalone file to decouple ViewModel from View
- Log error when AR carrier package selection finds no match
