+++ Modules/Sources/ParcelFittingCheck/ARUnifiedParcelFlowView.swift
@@ -4,7 +4,7 @@ struct ARUnifiedParcelFlowView: View {
     let unit: UnitLength
     let carriers: [ParcelPresetCarrier]
     let starredPackageIDs: Set<String>
-    weak var delegate: ParcelFittingDelegate?
+    var delegate: ParcelFittingDelegate?
     let onCancel: () -> Void
     let onConfirm: (ParcelFittingResult) -> Void
