+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/WooShipping Create Shipping Labels/WooShipping Package and Rate Selection/ARPackageContext.swift
@@ -1,0 +1,8 @@
+import ParcelFittingCheck
+
+struct ARPackageContext {
+    let measurement: ParcelDimensions
+    let carriers: [ParcelPresetCarrier]
+    let starredPackageIDs: Set<String>
+    let dimensionUnit: UnitLength
+}
