+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/WooShipping Create Shipping Labels/WooShipping Package and Rate Selection/WooShippingAddCustomPackageViewModel.swift
@@ -1,5 +1,3 @@
-import ARKit
-import Experiments
 import Foundation
 import Yosemite
 import protocol WooFoundation.Analytics
@@ -8,7 +6,6 @@ final class WooShippingAddCustomPackageViewModel: ObservableObject {
     private let stores: StoresManager
     private let siteID: Int64
     private let analytics: Analytics
-    private let featureFlagService: FeatureFlagService
 
     // Holds values for all dimension input fields.
     // Using a dictionary so we can easily add/remove new types
@@ -20,11 +17,6 @@ final class WooShippingAddCustomPackageViewModel: ObservableObject {
     @Published var showSaveTemplate: Bool = false
     @Published var packageTemplateName: String = ""
 
-    var isARParcelFittingAvailable: Bool {
-        featureFlagService.isFeatureFlagEnabled(.arParcelFitting)
-        && ARWorldTrackingConfiguration.isSupported
-    }
-
     var packageID: String {
         packageTemplateName.isEmpty ? Constants.defaultBoxID : packageTemplateName
     }
@@ -34,12 +26,10 @@ final class WooShippingAddCustomPackageViewModel: ObservableObject {
     init(selectedPackage: WooShippingPackageDataRepresentable? = nil,
          siteID: Int64 = ServiceLocator.stores.sessionManager.defaultStoreID ?? 0,
          stores: StoresManager = ServiceLocator.stores,
-         analytics: Analytics = ServiceLocator.analytics,
-         featureFlagService: FeatureFlagService = ServiceLocator.featureFlagService) {
+         analytics: Analytics = ServiceLocator.analytics) {
         self.stores = stores
         self.siteID = siteID
         self.analytics = analytics
-        self.featureFlagService = featureFlagService
         if let selectedPackage {
             fieldValues = [
                 .length: selectedPackage.length,
